refactor: Move the logic to load user's function to loader.ts
This separates the logic to load user's function and one to start the webserver. Moves the function signature type definition to function.ts. Also, adds more testcases for the function loading logic.

See also: #134
