+++ test/data/with_main/package.json
@@ -1,0 +1,3 @@
+{
+    "main": "foo.js"
+}
