+++ test/data/without_main/function.js
@@ -1,0 +1,24 @@
+/**
+ * 
+ *
+ * @param {!Object} req request context.
+ * @param {!Object} res response context.
+ */
+function testHttpFunction (res, req) {
+  return 'PASS'
+};
+
+/**
+ * 
+ *
+ * @param {!Object} data event payload.
+ * @param {!Object} context event metadata.
+ */
+function testEventFunction (data, context) {
+  return 'PASS';
+};
+
+module.exports = {
+  testHttpFunction,
+  testEventFunction,
+}
